+++ app/RELEASE_NOTES.md
@@ -1,3 +1,4 @@
 # Release notes
 * Now supports iPad.
 * Improves notification permission flow.
+* Adds Street View to MapCard (please test).